Moving primitive parameters to cell generator
