+++ align/primitive/__init__.py
@@ -1,1 +1,1 @@
-from .main import generate_primitive, get_generator
+from .main import generate_primitive, get_generator, generate_primitive_lef
